requirements.txt
cachelib==0.10.2
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
distlib==0.3.6
et-xmlfile==1.1.0
filelock==3.12.0
Flask==2.2.3
Flask-Excel==0.0.7
Flask-MySQLdb==1.0.1
Flask-Session==0.4.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lml==0.1.0
MarkupSafe==2.1.2
mysql-connector-python==8.0.33
mysqlclient==2.1.1
openpyxl==3.1.2
platformdirs==3.5.1
protobuf==3.20.3
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
pyexcel-xlsx==0.6.0
requests==2.31.0
stripe==5.4.0
texttable==1.6.7
urllib3==2.0.3
virtualenv==20.23.0
Werkzeug==2.2.3
